separate out classes for base query
